+++ src/parsing/parser.ml
@@ -15,16 +15,32 @@ open Common
 let parser_fatal : Pos.pos -> ('a,'b) koutfmt -> 'a = fun pos fmt ->
   Error.fatal (Some(pos)) fmt
 
-(** [qident_of_string s] converts the string [s] into a path. *)
-let qident_of_string : string -> (Syntax.qident, Pos.popt) result = fun s ->
-    let lb = Sedlexing.Utf8.from_string s in
-    let token = LpLexer.token lb in
-    let parse =
-      MenhirLib.Convert.Simplified.traditional2revised LpParser.id in
-    try Ok(let Pos.{elt;_} = parse token in elt)
-    with LpLexer.SyntaxError(s) -> Error(s.pos)
-       | LpParser.Error ->
-           Error(Some(Pos.locate (Sedlexing.lexing_positions lb)))
+(** [path_of_string s] converts the string [s] into a path. *)
+let path_of_string : string -> Path.t = fun s ->
+  let open LpLexer in
+  let lb = Sedlexing.Utf8.from_string s in
+  try
+    begin match token lb () with
+      | ID s, _, _ -> [s]
+      | PATH p, _, _ -> List.rev p
+      | _ -> Error.fatal_no_pos "Invalid path."
+    end
+  with SyntaxError {elt;_} ->
+    Error.fatal_no_pos "Syntax error in path: %s." elt
+
+(** [qident_of_string s] converts the string [s] into a qident. *)
+let qident_of_string : string -> Core.Term.qident = fun s ->
+  let open LpLexer in
+  let lb = Sedlexing.Utf8.from_string s in
+  try
+    begin match token lb () with
+      | ID s, _, _ -> ([], s)
+      | PATH [], _, _ -> assert false
+      | PATH (s::p), _, _ -> (List.rev p, s)
+      | _ -> Error.fatal_no_pos "Invalid qualified identifier."
+    end
+  with SyntaxError {elt;_} ->
+    Error.fatal_no_pos "Syntax error in qualified identifier: %s." elt
 
 (** Module type of a parser. *)
 module type PARSER = sig